+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **PODAAC-4360**
   - code change to be compliant with cumulus 11 message schema
   - move to public github and construct github action 
+  - enhanced the code so while replacing string, not only targeting fileName, key and source
+    but also consider that only key and bucket are required under new CMA message schema. 
+    i.e. fileName and source could be missing.
 
 ### Deprecated
 ### Removed